Adding dockerignore for openwrt build container
If youve already got a build cache from previous builds in `./build`
and try to the build the docker container for the openwrt build
environment it will end up trying to include the entire build cache
into the container. This will result in a container that 11G+ in size.

Rather than doing that well just ignore the `./build` and other
unnecessary containers during the docker build.